+++ src/main/java/Parser.java
@@ -1,20 +1,30 @@
 import command.*;
 import exception.MikiArgsException;
 import exception.TaskParseException;
-import storage.Storage;
 import task.Deadline;
 import task.Event;
 import task.Task;
 import task.Todo;
 
-import java.io.IOException;
 import java.time.LocalDateTime;
 import java.time.format.DateTimeParseException;
 import java.util.Arrays;
 import java.util.Collections;
 
+/**
+ * A parser for Miki interactive command-line inputs.
+ */
 public class Parser {
-    public static List parseList(String[] args) throws MikiArgsException {
+    /**
+     * Parses a <code>ListTasks</code> command from a provided <code>String[]</code> of
+     * space-separated tokens.
+     *
+     * @param args <code>String[]</code> denoting a <code>ListTasks</code> command
+     * @return a <code>ListTasks</code> command represented by <code>args</code>
+     * @throws MikiArgsException if <code>args</code> does not represent
+     * a valid <code>ListTasks</code> command
+     */
+    public static ListTasks parseList(String[] args) throws MikiArgsException {
         String from = "";
         String to = "";
         boolean token_from = false;
@@ -54,9 +64,16 @@ public class Parser {
                 throw new MikiArgsException(to + " needs to be formatted as " + Task.DATE_IN_FMT_STR + "!");
             }
         }
-        return new List(fromDate, toDate);
+        return new ListTasks(fromDate, toDate);
     }
 
+    /**
+     * Parses a <code>TaskList</code> index from a <code>String[]</code> of space-separated tokens.
+     *
+     * @param args <code>String[]</code> containing a <code>TaskList</code> index
+     * @return the <code>TaskList</code> index represented by <code>args</code>
+     * @throws MikiArgsException if <code>args</code> does not represent an integer
+     */
     public static int parseTaskIndex(String[] args) throws MikiArgsException {
         int idx;
         if (args.length == 0) {
@@ -70,6 +87,13 @@ public class Parser {
         return idx;
     }
 
+    /**
+     * Forms a single <code>String</code> representing a file path by combining
+     * consecutive fragmented <code>String</code>s in a <code>String[]</code> of space-separated tokens.
+     *
+     * @param args fragmented <code>String[]</code> to combine
+     * @return the file path represented by <code>args</code>
+     */
     public static String parsePath(String[] args) {
         String path = "";
         for (int i = 0; i < args.length; i++) {
@@ -78,10 +102,22 @@ public class Parser {
         return path;
     }
 
+    /**
+     * Returns <code>true</code> if the supplied line of input is requesting the program to exit.
+     *
+     * @param cmdLine the input to parse
+     * @return <code>true</code> if <code>cmdLine</code> represents an exit command
+     */
     public static boolean parseExit(String cmdLine) {
         return cmdLine.split(" ")[0].toLowerCase().equals("bye");
     }
 
+    /**
+     * Parses one supplied line of input String to create a <code>Command</code> action.
+     *
+     * @param cmdLine the input denoting a <code>Command</code>
+     * @return a <code>Command</code> represented by <code>cmdLine</code>
+     */
     public static Command parse(String cmdLine) {
         String cmd = cmdLine.split(" ")[0].toLowerCase();
         String[] args = {};
@@ -91,7 +127,7 @@ public class Parser {
         try {
             switch (cmd) {
                 case "bye":
-                    return new Exit();
+                    return new ExitPrint();
                 case "list":
                     return parseList(args);
                 case "mark":
