+++ src/main/java/command/Delete.java
@@ -6,13 +6,28 @@ import storage.Storage;
 import task.Task;
 import task.TaskList;
 
+/**
+ * A command deleting a <code>Task</code> at a specified index of a <code>TaskList</code>.
+ */
 public class Delete implements Command {
     private final int idx;
 
+    /**
+     * Creates a delete-task command.
+     *
+     * @param idx index to be deleted
+     */
     public Delete(int idx) {
         this.idx = idx;
     }
 
+    /**
+     * {@inheritDoc}
+     *
+     * @param tasks tasklist to perform the action on
+     * @param ui ui to perform the action on
+     * @param storage storage to perform the action on
+     */
     @Override
     public void run(TaskList tasks, Ui ui, Storage storage) {
         try {
